Move helper functions to other file, update README
